refactor: package qna modules
