+++ app.py
@@ -1,9 +1,15 @@
 import streamlit as st
 import json
 import os
-from data_processor import extract_text_from_url, extract_text_from_uploaded_file
-from ai_qa_generator import AIQAGenerator
-from data_exporter import export_to_jsonl, export_to_json, export_to_csv, export_for_rag, export_for_finetuning
+from qna_generator.data_processor import extract_text_from_url, extract_text_from_uploaded_file
+from qna_generator.ai_qa_generator import AIQAGenerator
+from qna_generator.data_exporter import (
+    export_to_jsonl,
+    export_to_json,
+    export_to_csv,
+    export_for_rag,
+    export_for_finetuning,
+)
 
 # ページ設定
 st.set_page_config(
