--- qna_generator/qna_generator/README.md
@@ -1,55 +1,0 @@
-# Q&A自動生成システム
-
-URL、PDF、DOCXファイルからQ&Aを自動生成するStreamlitアプリケーションです。
-
-## 機能
-
-- **データ入力**: URL、PDF、DOCXファイルからテキストを抽出
-- **AI Q&A生成**: GPT-4.1 nanoを使用してカテゴリ別Q&Aを自動生成
-- **温度調整**: AIモデルの創造性を0.0〜0.8で調整可能
-- **引用元表示**: 回答の引用元を折りたたみ表示で確認可能
-- **データエクスポート**: JSON、CSV、JSONL、RAG用、ファインチューニング用の各形式でダウンロード
-
-## 必要な環境
-
-- Python 3.11+
-- OpenAI APIキー
-
-## インストール
-
-1. 必要なライブラリをインストール:
-```bash
-pip install -r requirements.txt
-```
-
-2. アプリケーションを起動:
-```bash
-streamlit run app.py
-```
-
-## 使用方法
-
-1. **APIキー設定**: サイドバーでOpenAI APIキーを入力
-2. **データ入力**: URLまたはファイルをアップロードしてテキストを抽出
-3. **Q&A生成**: 「カテゴリとQ&Aを生成」ボタンをクリック
-4. **結果確認**: 生成されたQ&Aを確認し、引用元を展開して詳細を確認
-5. **データエクスポート**: 必要に応じて各種形式でデータをダウンロード
-
-## ファイル構成
-
-- `app.py`: メインアプリケーション
-- `data_processor.py`: データ抽出モジュール
-- `ai_qa_generator.py`: AI Q&A生成エンジン
-- `data_exporter.py`: データエクスポート機能
-- `requirements.txt`: 必要なライブラリ一覧
-
-## 注意事項
-
-- OpenAI APIキーが必要です
-- 大きなファイルの処理には時間がかかる場合があります
-- 生成されるQ&Aの品質は入力データの質に依存します
-
-## ライセンス
-
-MIT License
-
