--- qna_generator/qna_generator/requirements.txt
@@ -1,8 +1,0 @@
-streamlit
-requests
-beautifulsoup4
-pymupdf
-python-docx
-openai
-
-
